geo API, авторизация, прочие штуки

+++ SpringServer/src/main/java/com/happytails/springserver/dto/PetDTO.java
@@ -24,7 +24,7 @@ public class PetDTO {
     @Positive
     Integer age;
     @Positive
-    Integer weight;
+    Double weight;
     String breed;
     String description;
     String photoPath;
